refactor: rename data.tensorflow to data.phasespace
